+++ src/main/java/org/nrg/containers/rest/CommandConfigurationRestApi.java
@@ -11,16 +11,12 @@ import org.nrg.containers.exceptions.NoDockerServerException;
 import org.nrg.containers.model.configuration.CommandConfiguration;
 import org.nrg.containers.model.configuration.ProjectEnabledReport;
 import org.nrg.containers.services.CommandService;
-import org.nrg.containers.services.ContainerConfigService;
 import org.nrg.containers.services.ContainerConfigService.CommandConfigurationException;
 import org.nrg.framework.annotations.XapiRestController;
 import org.nrg.framework.exceptions.NotFoundException;
 import org.nrg.xapi.rest.AbstractXapiRestController;
 import org.nrg.xapi.rest.XapiRequestMapping;
 import org.nrg.xdat.XDAT;
-import org.nrg.xdat.om.XnatProjectdata;
-import org.nrg.xdat.om.base.auto.AutoXnatProjectdata;
-import org.nrg.xdat.security.helpers.Permissions;
 import org.nrg.xdat.security.services.RoleHolder;
 import org.nrg.xdat.security.services.UserManagementServiceI;
 import org.nrg.xft.security.UserI;
@@ -28,7 +24,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -37,6 +32,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
+import static org.nrg.xdat.security.helpers.AccessLevel.Admin;
+import static org.nrg.xdat.security.helpers.AccessLevel.Owner;
+import static org.nrg.xdat.security.helpers.AccessLevel.Read;
 import static org.springframework.web.bind.annotation.RequestMethod.DELETE;
 import static org.springframework.web.bind.annotation.RequestMethod.GET;
 import static org.springframework.web.bind.annotation.RequestMethod.POST;
@@ -47,28 +45,18 @@ import static org.springframework.web.bind.annotation.RequestMethod.PUT;
 public class CommandConfigurationRestApi extends AbstractXapiRestController {
     private static final Logger log = LoggerFactory.getLogger(CommandConfigurationRestApi.class);
 
-    private static final String JSON = MediaType.APPLICATION_JSON_UTF8_VALUE;
-    private static final String TEXT = MediaType.TEXT_PLAIN_VALUE;
-    private static final String FORM = MediaType.APPLICATION_FORM_URLENCODED_VALUE;
-
-    private static final String ID_REGEX = "\\d+";
-    private static final String NAME_REGEX = "\\d*[^\\d]+\\d*";
-
     private CommandService commandService;
-    private ContainerConfigService containerConfigService;
 
     @Autowired
     public CommandConfigurationRestApi(final CommandService commandService,
-                                       final ContainerConfigService containerConfigService,
                                        final UserManagementServiceI userManagementService,
                                        final RoleHolder roleHolder) {
         super(userManagementService, roleHolder);
         this.commandService = commandService;
-        this.containerConfigService = containerConfigService;
     }
 
     // Configure for site + command wrapper
-    @XapiRequestMapping(value = {"/commands/{commandId}/wrappers/{wrapperName}/config"}, method = POST)
+    @XapiRequestMapping(value = {"/commands/{commandId}/wrappers/{wrapperName}/config"}, method = POST, restrictTo = Admin)
     public ResponseEntity<Void> createConfiguration(final @RequestBody CommandConfiguration commandConfiguration,
                                                     final @PathVariable long commandId,
                                                     final @PathVariable String wrapperName,
@@ -76,79 +64,57 @@ public class CommandConfigurationRestApi extends AbstractXapiRestController {
                                                     final @RequestParam(required = false) String reason)
             throws CommandConfigurationException, NotFoundException {
         final UserI userI = XDAT.getUserDetails();
-        final HttpStatus status = isAdminUser();
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
+
         commandService.configureForSite(commandConfiguration, commandId, wrapperName, enable, userI.getLogin(), reason);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
-    @XapiRequestMapping(value = {"/wrappers/{wrapperId}/config"}, method = POST)
+    @XapiRequestMapping(value = {"/wrappers/{wrapperId}/config"}, method = POST, restrictTo = Admin)
     public ResponseEntity<Void> createConfiguration(final @RequestBody CommandConfiguration commandConfiguration,
                                                     final @PathVariable long wrapperId,
                                                     final @RequestParam(required = false, defaultValue = "true") boolean enable,
                                                     final @RequestParam(required = false) String reason)
             throws CommandConfigurationException, NotFoundException {
         final UserI userI = XDAT.getUserDetails();
-        final HttpStatus status = isAdminUser();
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
+
         commandService.configureForSite(commandConfiguration, wrapperId, enable, userI.getLogin(), reason);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     // Get configuration for site + command wrapper
-    @XapiRequestMapping(value = {"/commands/{commandId}/wrappers/{wrapperName}/config"}, method = GET)
+    @XapiRequestMapping(value = {"/commands/{commandId}/wrappers/{wrapperName}/config"}, method = GET, restrictTo = Admin)
     @ResponseBody
     public CommandConfiguration getConfiguration(final @PathVariable long commandId,
                                                  final @PathVariable String wrapperName) throws NotFoundException {
-        final HttpStatus status = isAdminUser();
-        if (status != null) {
-            return null;
-        }
         return commandService.getSiteConfiguration(commandId, wrapperName);
     }
 
-    @XapiRequestMapping(value = {"/wrappers/{wrapperId}/config"}, method = GET)
+    @XapiRequestMapping(value = {"/wrappers/{wrapperId}/config"}, method = GET, restrictTo = Admin)
     @ResponseBody
     public CommandConfiguration getConfiguration(final @PathVariable long wrapperId) throws NotFoundException {
-        final HttpStatus status = isAdminUser();
-        if (status != null) {
-            return null;
-        }
         return commandService.getSiteConfiguration(wrapperId);
     }
 
     // Delete configuration for site + command wrapper
-    @XapiRequestMapping(value = {"/commands/{commandId}/wrappers/{wrapperName}/config"}, method = DELETE)
+    @XapiRequestMapping(value = {"/commands/{commandId}/wrappers/{wrapperName}/config"}, method = DELETE, restrictTo = Admin)
     public ResponseEntity<Void> deleteConfiguration(final @PathVariable long commandId,
                                                     final @PathVariable String wrapperName)
             throws CommandConfigurationException, NotFoundException {
-        final HttpStatus status = isAdminUser();
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
         final UserI userI = XDAT.getUserDetails();
         commandService.deleteSiteConfiguration(commandId, wrapperName, userI.getLogin());
         return ResponseEntity.noContent().build();
     }
 
-    @XapiRequestMapping(value = {"/wrappers/{wrapperId}/config"}, method = DELETE)
+    @XapiRequestMapping(value = {"/wrappers/{wrapperId}/config"}, method = DELETE, restrictTo = Admin)
     public ResponseEntity<Void> deleteConfiguration(final @PathVariable long wrapperId)
             throws CommandConfigurationException, NotFoundException {
-        final HttpStatus status = isAdminUser();
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
         final UserI userI = XDAT.getUserDetails();
         commandService.deleteSiteConfiguration(wrapperId, userI.getLogin());
         return ResponseEntity.noContent().build();
     }
 
     // Configure for project + command wrapper
-    @XapiRequestMapping(value = {"/projects/{project}/commands/{commandId}/wrappers/{wrapperName}/config"}, method = POST)
+    @XapiRequestMapping(value = {"/projects/{project}/commands/{commandId}/wrappers/{wrapperName}/config"}, method = POST, restrictTo = Owner)
     public ResponseEntity<Void> createConfiguration(final @RequestBody CommandConfiguration commandConfiguration,
                                                     final @PathVariable String project,
                                                     final @PathVariable long commandId,
@@ -157,15 +123,12 @@ public class CommandConfigurationRestApi extends AbstractXapiRestController {
                                                     final @RequestParam(required = false) String reason)
             throws CommandConfigurationException, NotFoundException {
         final UserI userI = XDAT.getUserDetails();
-        final HttpStatus status = isProjectOwnerOrAdmin(project);
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
+
         commandService.configureForProject(commandConfiguration, project, commandId, wrapperName, enable, userI.getLogin(), reason);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
-    @XapiRequestMapping(value = {"/projects/{project}/wrappers/{wrapperId}/config"}, method = POST)
+    @XapiRequestMapping(value = {"/projects/{project}/wrappers/{wrapperId}/config"}, method = POST, restrictTo = Owner)
     public ResponseEntity<Void> createConfiguration(final @RequestBody CommandConfiguration commandConfiguration,
                                                     final @PathVariable String project,
                                                     final @PathVariable long wrapperId,
@@ -173,61 +136,42 @@ public class CommandConfigurationRestApi extends AbstractXapiRestController {
                                                     final @RequestParam(required = false) String reason)
             throws CommandConfigurationException, NotFoundException {
         final UserI userI = XDAT.getUserDetails();
-        final HttpStatus status = isProjectOwnerOrAdmin(project);
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
+
         commandService.configureForProject(commandConfiguration, project, wrapperId, enable, userI.getLogin(), reason);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     // Get configuration for project + command wrapper
-    @XapiRequestMapping(value = {"/projects/{project}/commands/{commandId}/wrappers/{wrapperName}/config"}, method = GET)
+    @XapiRequestMapping(value = {"/projects/{project}/commands/{commandId}/wrappers/{wrapperName}/config"}, method = GET, restrictTo = Read)
     @ResponseBody
     public CommandConfiguration getConfiguration(final @PathVariable String project,
                                                  final @PathVariable long commandId,
                                                  final @PathVariable String wrapperName) throws NotFoundException {
-        final HttpStatus status = canReadProjectOrAdmin(project);
-        if (status != null) {
-            return null;
-        }
         return commandService.getProjectConfiguration(project, commandId, wrapperName);
     }
 
-    @XapiRequestMapping(value = {"/projects/{project}/wrappers/{wrapperId}/config"}, method = GET)
+    @XapiRequestMapping(value = {"/projects/{project}/wrappers/{wrapperId}/config"}, method = GET, restrictTo = Read)
     @ResponseBody
     public CommandConfiguration getConfiguration(final @PathVariable String project,
                                                  final @PathVariable long wrapperId) throws NotFoundException {
-        final HttpStatus status = canReadProjectOrAdmin(project);
-        if (status != null) {
-            return null;
-        }
         return commandService.getProjectConfiguration(project, wrapperId);
     }
 
     // Delete configuration for project + command wrapper
-    @XapiRequestMapping(value = {"/projects/{project}/commands/{commandId}/wrappers/{wrapperName}/config"}, method = DELETE)
+    @XapiRequestMapping(value = {"/projects/{project}/commands/{commandId}/wrappers/{wrapperName}/config"}, method = DELETE, restrictTo = Owner)
     public ResponseEntity<Void> deleteConfiguration(final @PathVariable String project,
                                                     final @PathVariable long commandId,
                                                     final @PathVariable String wrapperName)
             throws CommandConfigurationException, NotFoundException {
-        final HttpStatus status = isProjectOwnerOrAdmin(project);
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
         final UserI userI = XDAT.getUserDetails();
         commandService.deleteProjectConfiguration(project, commandId, wrapperName, userI.getLogin());
         return ResponseEntity.noContent().build();
     }
 
-    @XapiRequestMapping(value = {"/projects/{project}/wrappers/{wrapperId}/config"}, method = DELETE)
+    @XapiRequestMapping(value = {"/projects/{project}/wrappers/{wrapperId}/config"}, method = DELETE, restrictTo = Owner)
     public ResponseEntity<Void> deleteConfiguration(final @PathVariable String project,
                                                     final @PathVariable long wrapperId)
             throws CommandConfigurationException, NotFoundException {
-        final HttpStatus status = isProjectOwnerOrAdmin(project);
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
         final UserI userI = XDAT.getUserDetails();
         commandService.deleteProjectConfiguration(project, wrapperId, userI.getLogin());
         return ResponseEntity.noContent().build();
@@ -236,161 +180,113 @@ public class CommandConfigurationRestApi extends AbstractXapiRestController {
     /*
     ENABLE/DISABLE
      */
-    @XapiRequestMapping(value = {"/commands/{commandId}/wrappers/{wrapperName}/enabled"}, method = GET)
+    @XapiRequestMapping(value = {"/commands/{commandId}/wrappers/{wrapperName}/enabled"}, method = GET, restrictTo = Admin)
     @ResponseBody
     public Boolean isConfigurationEnabled(final @PathVariable long commandId,
                                           final @PathVariable String wrapperName)
             throws CommandConfigurationException, NotFoundException {
-        final HttpStatus status = isAdminUser();
-        if (status != null) {
-            return false;
-        }
         return commandService.isEnabledForSite(commandId, wrapperName);
     }
 
-    @XapiRequestMapping(value = {"/wrappers/{wrapperId}/enabled"}, method = GET)
+    @XapiRequestMapping(value = {"/wrappers/{wrapperId}/enabled"}, method = GET, restrictTo = Admin)
     @ResponseBody
     public Boolean isConfigurationEnabled(final @PathVariable long wrapperId)
             throws CommandConfigurationException, NotFoundException {
-        final HttpStatus status = isAdminUser();
-        if (status != null) {
-            return false;
-        }
         return commandService.isEnabledForSite(wrapperId);
     }
 
-    @XapiRequestMapping(value = {"/commands/{commandId}/wrappers/{wrapperName}/enabled"}, method = PUT)
+    @XapiRequestMapping(value = {"/commands/{commandId}/wrappers/{wrapperName}/enabled"}, method = PUT, restrictTo = Admin)
     public ResponseEntity<Void> enableConfiguration(final @PathVariable long commandId,
                                                     final @PathVariable String wrapperName,
                                                     final @RequestParam(required = false) String reason)
             throws CommandConfigurationException, NotFoundException {
         final UserI userI = XDAT.getUserDetails();
-        final HttpStatus status = isAdminUser();
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
         commandService.enableForSite(commandId, wrapperName, userI.getLogin(), reason);
         return ResponseEntity.ok().build();
     }
 
-    @XapiRequestMapping(value = {"/wrappers/{wrapperId}/enabled"}, method = PUT)
+    @XapiRequestMapping(value = {"/wrappers/{wrapperId}/enabled"}, method = PUT, restrictTo = Admin)
     public ResponseEntity<Void> enableConfiguration(final @PathVariable long wrapperId,
                                                     final @RequestParam(required = false) String reason)
             throws CommandConfigurationException, NotFoundException {
         final UserI userI = XDAT.getUserDetails();
-        final HttpStatus status = isAdminUser();
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
         commandService.enableForSite(wrapperId, userI.getLogin(), reason);
         return ResponseEntity.ok().build();
     }
 
-    @XapiRequestMapping(value = {"/commands/{commandId}/wrappers/{wrapperName}/disabled"}, method = PUT)
+    @XapiRequestMapping(value = {"/commands/{commandId}/wrappers/{wrapperName}/disabled"}, method = PUT, restrictTo = Admin)
     public ResponseEntity<Void> disableConfiguration(final @PathVariable long commandId,
                                                      final @PathVariable String wrapperName,
                                                      final @RequestParam(required = false) String reason)
             throws CommandConfigurationException, NotFoundException {
         final UserI userI = XDAT.getUserDetails();
-        final HttpStatus status = isAdminUser();
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
         commandService.disableForSite(commandId, wrapperName, userI.getLogin(), reason);
         return ResponseEntity.ok().build();
     }
 
-    @XapiRequestMapping(value = {"/wrappers/{wrapperId}/disabled"}, method = PUT)
+    @XapiRequestMapping(value = {"/wrappers/{wrapperId}/disabled"}, method = PUT, restrictTo = Admin)
     public ResponseEntity<Void> disableConfiguration(final @PathVariable long wrapperId,
                                                      final @RequestParam(required = false) String reason)
             throws CommandConfigurationException, NotFoundException {
         final UserI userI = XDAT.getUserDetails();
-        final HttpStatus status = isAdminUser();
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
         commandService.disableForSite(wrapperId, userI.getLogin(), reason);
         return ResponseEntity.ok().build();
     }
 
-    @XapiRequestMapping(value = {"/projects/{project}/commands/{commandId}/wrappers/{wrapperName}/enabled"}, method = GET)
-    public ResponseEntity<ProjectEnabledReport> isConfigurationEnabled(final @PathVariable String project,
+    @XapiRequestMapping(value = {"/projects/{project}/commands/{commandId}/wrappers/{wrapperName}/enabled"}, method = GET, restrictTo = Read)
+    public ProjectEnabledReport isConfigurationEnabled(final @PathVariable String project,
                                                                        final @PathVariable long commandId,
                                                                        final @PathVariable String wrapperName)
             throws CommandConfigurationException, NotFoundException {
-        final HttpStatus status = canReadProjectOrAdmin(project);
-        if (status != null) {
-            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
-        }
-        return ResponseEntity.ok(commandService.isEnabledForProjectAsReport(project, commandId, wrapperName));
+        return commandService.isEnabledForProjectAsReport(project, commandId, wrapperName);
     }
 
-    @XapiRequestMapping(value = {"/projects/{project}/wrappers/{wrapperId}/enabled"}, method = GET)
+    @XapiRequestMapping(value = {"/projects/{project}/wrappers/{wrapperId}/enabled"}, method = GET, restrictTo = Read)
     @ResponseBody
-    public ResponseEntity<ProjectEnabledReport> isConfigurationEnabled(final @PathVariable String project,
+    public ProjectEnabledReport isConfigurationEnabled(final @PathVariable String project,
                                                                        final @PathVariable long wrapperId)
             throws CommandConfigurationException, NotFoundException {
-        final HttpStatus status = canReadProjectOrAdmin(project);
-        if (status != null) {
-            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
-        }
-        return ResponseEntity.ok(commandService.isEnabledForProjectAsReport(project, wrapperId));
+        return commandService.isEnabledForProjectAsReport(project, wrapperId);
     }
 
-    @XapiRequestMapping(value = {"/projects/{project}/commands/{commandId}/wrappers/{wrapperName}/enabled"}, method = PUT)
+    @XapiRequestMapping(value = {"/projects/{project}/commands/{commandId}/wrappers/{wrapperName}/enabled"}, method = PUT, restrictTo = Owner)
     public ResponseEntity<Void> enableConfiguration(final @PathVariable String project,
                                                     final @PathVariable long commandId,
                                                     final @PathVariable String wrapperName,
                                                     final @RequestParam(required = false) String reason)
             throws CommandConfigurationException, NotFoundException {
         final UserI userI = XDAT.getUserDetails();
-        final HttpStatus status = isProjectOwnerOrAdmin(project);
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
         commandService.enableForProject(project, commandId, wrapperName, userI.getLogin(), reason);
         return ResponseEntity.ok().build();
     }
 
-    @XapiRequestMapping(value = {"/projects/{project}/wrappers/{wrapperId}/enabled"}, method = PUT)
+    @XapiRequestMapping(value = {"/projects/{project}/wrappers/{wrapperId}/enabled"}, method = PUT, restrictTo = Owner)
     public ResponseEntity<Void> enableConfiguration(final @PathVariable String project,
                                                     final @PathVariable long wrapperId,
                                                     final @RequestParam(required = false) String reason)
             throws CommandConfigurationException, NotFoundException {
         final UserI userI = XDAT.getUserDetails();
-        final HttpStatus status = isProjectOwnerOrAdmin(project);
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
         commandService.enableForProject(project, wrapperId, userI.getLogin(), reason);
         return ResponseEntity.ok().build();
     }
 
-    @XapiRequestMapping(value = {"/projects/{project}/commands/{commandId}/wrappers/{wrapperName}/disabled"}, method = PUT)
+    @XapiRequestMapping(value = {"/projects/{project}/commands/{commandId}/wrappers/{wrapperName}/disabled"}, method = PUT, restrictTo = Owner)
     public ResponseEntity<Void> disableConfiguration(final @PathVariable String project,
                                                      final @PathVariable long commandId,
                                                      final @PathVariable String wrapperName,
                                                      final @RequestParam(required = false) String reason)
             throws CommandConfigurationException, NotFoundException {
         final UserI userI = XDAT.getUserDetails();
-        final HttpStatus status = isProjectOwnerOrAdmin(project);
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
         commandService.disableForProject(project, commandId, wrapperName, userI.getLogin(), reason);
         return ResponseEntity.ok().build();
     }
 
-    @XapiRequestMapping(value = {"/projects/{project}/wrappers/{wrapperId}/disabled"}, method = PUT)
+    @XapiRequestMapping(value = {"/projects/{project}/wrappers/{wrapperId}/disabled"}, method = PUT, restrictTo = Owner)
     public ResponseEntity<Void> disableConfiguration(final @PathVariable String project,
                                                      final @PathVariable long wrapperId,
                                                      final @RequestParam(required = false) String reason)
             throws CommandConfigurationException, NotFoundException {
         final UserI userI = XDAT.getUserDetails();
-        final HttpStatus status = isProjectOwnerOrAdmin(project);
-        if (status != null) {
-            return new ResponseEntity<>(status);
-        }
         commandService.disableForProject(project, wrapperId, userI.getLogin(), reason);
         return ResponseEntity.ok().build();
     }
@@ -456,61 +352,5 @@ public class CommandConfigurationRestApi extends AbstractXapiRestController {
         }
         log.debug(message);
         return message;
-    }
-
-    /**
-     * Checks if is permitted.
-     *
-     * @param projectId the project ID
-     *
-     * @return the http status
-     */
-    // TODO: Migrate this to the abstract superclass. Can't right now because XDAT doesn't know about XnatProjectdata, etc.
-    private HttpStatus canReadProjectOrAdmin(String projectId) {
-        final UserI sessionUser = getSessionUser();
-        if (projectId != null) {
-            final XnatProjectdata project = AutoXnatProjectdata.getXnatProjectdatasById(projectId, sessionUser, false);
-            try {
-                return ( Permissions.canRead(sessionUser, project) || getRoleHolder().isSiteAdmin(sessionUser) ) ? null : HttpStatus.FORBIDDEN;
-            } catch (Exception e) {
-                log.error("Error checking read status for project", e);
-                return HttpStatus.INTERNAL_SERVER_ERROR;
-            }
-        } else {
-            return isPermitted() == null ? null : HttpStatus.FORBIDDEN;
-        }
-    }
-
-    /**
-     * Checks if user is owner.
-     *
-     * @param projectId the project ID
-     *
-     * @return the http status
-     */
-    // TODO: Migrate this to the abstract superclass. Can't right now because XDAT doesn't know about XnatProjectdata, etc.
-    private HttpStatus isProjectOwnerOrAdmin(String projectId) {
-        final UserI sessionUser = getSessionUser();
-        if (projectId != null) {
-            final XnatProjectdata project = AutoXnatProjectdata.getXnatProjectdatasById(projectId, sessionUser, false);
-            try {
-                return ( Permissions.isProjectOwner(sessionUser, projectId) || getRoleHolder().isSiteAdmin(sessionUser) ) ? null : HttpStatus.FORBIDDEN;
-            } catch (Exception e) {
-                log.error("Error checking read status for project", e);
-                return HttpStatus.INTERNAL_SERVER_ERROR;
-            }
-        } else {
-            return isPermitted() == null ? null : HttpStatus.FORBIDDEN;
-        }
-    }
-
-    private HttpStatus isAdminUser() {
-        final UserI sessionUser = getSessionUser();
-        try {
-            return getRoleHolder().isSiteAdmin(sessionUser) ? null : HttpStatus.FORBIDDEN;
-        } catch (Exception e) {
-            log.error("Error checking whether admin is project owner", e);
-            return HttpStatus.INTERNAL_SERVER_ERROR;
-        }
     }
 }